% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat.uploadResponse}
\alias{chat.uploadResponse}
\title{Upload Chat Response to Database}
\usage{
chat.uploadResponse(response, con)
}
\arguments{
\item{response}{A list containing the response from the OpenAI Chat API.
The list structure should include \code{chatID}, a nested \code{results} list with \code{choices} and \code{model}.}

\item{con}{A PostgreSQL connection object created using DBI.
This connection is used to upload the data.}
}
\value{
Invisible NULL; the function is used for its side effects of uploading data.
}
\description{
This function takes a response from the OpenAI Chat API and uploads relevant information to a PostgreSQL database.
It extracts the chat ID, response content, model used for the response, and the timestamp of the upload, then
packages these into a data frame which is uploaded to a specified database table.
}
\examples{
\dontrun{
  # Assuming `response` is a response object from `chat.openAI` and `con` is a valid DBI connection
  chat.uploadResponse(response, con)
}
}
\seealso{
\code{\link[PDaiPostgres]{postgres.uploadData}}
}
