% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets.aggregate}
\alias{datasets.aggregate}
\title{Aggregate Large Datasets}
\usage{
datasets.aggregate(datasets, threshold = 100)
}
\arguments{
\item{datasets}{A list of data frames, each representing a dataset to potentially be aggregated.}

\item{threshold}{An integer, the minimum number of rows a dataset must have to be considered for aggregation.}
}
\value{
A list of datasets where datasets exceeding the specified threshold are aggregated.
}
\description{
This function iterates over a list of datasets and aggregates them based on a specified row threshold.
It is particularly useful for reducing larger datasets into a more manageable size by averaging over specific intervals.
}
\examples{
\dontrun{
  data_list <- list(dataset1 = data.frame(date = as.Date('2020-01-01') + 0:150, value = rnorm(151)),
                    dataset2 = data.frame(date = as.Date('2020-01-01') + 0:50, value = rnorm(51)))
  aggregated_data <- datasets.aggregate(data_list, threshold = 100)
  print(aggregated_data)
}
}
