% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{ts.aggregate}
\alias{ts.aggregate}
\title{Time Series Aggregation Function}
\usage{
ts.aggregate(
  ts,
  date_name = "ds",
  includeColumns = c("trend"),
  outputLength = 100,
  metric = "mean"
)
}
\arguments{
\item{ts}{A data frame or data.table containing the time series data.}

\item{date_name}{The name of the column in \code{ts} that contains the date information (defaults to 'ds').}

\item{includeColumns}{A character vector specifying which columns to include in the aggregation (defaults to 'trend').}

\item{outputLength}{The desired number of output groups (defaults to 100).}

\item{metric}{The aggregation metric to apply ('mean', 'median', 'mode', or 'weighted').}
}
\value{
A list containing a description of the aggregation and a data.table summarizing the aggregated data.
}
\description{
This function aggregates a time series data set based on a specified metric. It converts the data into
a data.table format, groups the data by calculated group sizes to ensure full coverage across the
specified output length, and applies a chosen aggregation function to the grouped data.
}
\examples{
\dontrun{
ts_data <- data.table(ds = seq(as.Date("2020/1/1"), by = "day", length.out = 300),
                      trend = rnorm(300))
aggregated_data <- ts.aggregate(ts_data, date_name = 'ds', includeColumns = c('trend'), outputLength = 10, metric = "mean")
print(aggregated_data)
}
}
